requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Pillow>=10.2.0
pdf2image==1.16.3
docx2pdf==0.1.8
reportlab==4.0.4
python-pptx==0.6.21
openpyxl==3.1.2
comtypes==1.2.0